fix: prevent possible slow resolution of rate limiter
We have found scenarios at scale where our requests lock up for 10+ seconds and wait for something in the rate limiter
while resolving a jwk endpoint.
We believe the intention is for the rate limiter to return immediately and not pause to wait for available tokens
(hence passing the fireImmediately arg to the RateLimiter constructor).

However, maybe because of a race condition at scale, it seems possible to end up trying
to resolve a token from the bucket even when there are no tokens left:
https://github.com/jhurliman/node-rate-limiter/blob/main/src/RateLimiter.ts#L83
Once we are on this path the rate limiter can trigger a wait for new 'tokens' become available:
https://github.com/jhurliman/node-rate-limiter/blob/main/src/TokenBucket.ts#L96

This fix uses the simpler `tryRemoveTokens` method which synchronously returns a boolean if tokens could be taken
(and so can't pause execution to wait for tokens to be available).